refactor(frame): rename main module to render module
